refactored some common code out of ds-api into common
