--- apps/ds-api/src/common/http-exception.ts
@@ -1,14 +1,0 @@
-export default class HttpException extends Error {
-  statusCode?: number;
-  status?: number;
-  message: string;
-  error: string | null;
-
-  constructor(statusCode: number, message: string, error?: string) {
-    super(message);
-
-    this.statusCode = statusCode;
-    this.message = message;
-    this.error = error || null;
-  }
-}
