--- apps/ds-api/src/middleware/error.middleware.ts
@@ -1,8 +1,0 @@
-import HttpException from '../common/http-exception';
-import { Request, Response, NextFunction } from 'express';
-
-export const errorHandler = (error: HttpException, request: Request, response: Response, next: NextFunction) => {
-  const status = error.statusCode || 500;
-
-  response.status(status).send(error);
-};
