+++ apps/ds-api/src/routes/actions/byId.ts
@@ -2,7 +2,7 @@ import express, { Request, Response } from 'express';
 import { ActionsForDay } from '@devouringscripture/common';
 import { db } from '../../services/db';
 import { param } from 'express-validator';
-import { validateRequest } from '../../middleware/validateRequest';
+import { validateRequest } from '@devouringscripture/common';
 
 const router = express.Router();
