+++ apps/ds-api/src/routes/actions/getStats.ts
@@ -1,8 +1,7 @@
 import express, { Request, Response } from 'express';
 import { query } from 'express-validator';
-import { validateRequest } from '../../middleware/validateRequest';
 import { db } from '../../services/db';
-import { ActionStats, ActionsForDay } from '@devouringscripture/common';
+import { ActionStats, ActionsForDay, validateRequest } from '@devouringscripture/common';
 import { DateTime } from 'luxon';
 
 const router = express.Router();
