+++ apps/ds-api/src/routes/actions/markItemForDate.ts
@@ -2,7 +2,7 @@ import express, { Request, Response } from 'express';
 import { db } from '../../services/db';
 import { getActionByIdInternal } from './byId';
 import { body, param } from 'express-validator';
-import { validateRequest } from '../../middleware/validateRequest';
+import { validateRequest } from '@devouringscripture/common';
 
 const router = express.Router();
