+++ apps/ds-api/src/routes/read/passages/delete.ts
@@ -1,6 +1,6 @@
 import express, { Request, Response } from 'express';
 import { param } from 'express-validator';
-import { validateRequest } from '../../../middleware/validateRequest';
+import { validateRequest } from '@devouringscripture/common';
 import { db } from '../../../services/db';
 
 const router = express.Router();
