+++ apps/ds-api/src/routes/read/passages/newItem.ts
@@ -1,7 +1,6 @@
 import express, { Request, Response } from 'express';
 import { body } from 'express-validator';
-import { validateRequest } from '../../../middleware/validateRequest';
-import { BasePassage, Passage } from '@devouringscripture/common';
+import { BasePassage, Passage, validateRequest } from '@devouringscripture/common';
 import { v4 as uuidv4 } from 'uuid';
 import { DateTime } from 'luxon';
 import { db } from '../../../services/db';
