+++ apps/ds-api/src/routes/user/update.ts
@@ -1,8 +1,7 @@
 import express, { Request, Response } from 'express';
 import { db } from '../../services/db';
-import { UserAttributes } from '@devouringscripture/common';
+import { UserAttributes, validateRequest } from '@devouringscripture/common';
 import { body, param } from 'express-validator';
-import { validateRequest } from '../../middleware/validateRequest';
 
 const router = express.Router();
