+++ packages/common/src/errors/CustomError.ts
