+++ packages/common/src/errors/RequestValidationError.ts
@@ -5,7 +5,7 @@ export class RequestValidationError extends CustomError {
   statusCode = 400;
 
   constructor(public errors: ValidationError[]) {
-    super('Validation error(s)');
+    super('Validation errors');
 
     Object.setPrototypeOf(this, RequestValidationError.prototype);
   }
