+++ packages/common/src/index.ts
@@ -3,3 +3,10 @@ export * from './dm/Passage';
 export * from './dm/PrayerTypes';
 export * from './dm/Prayer';
 export * from './dm/User';
+
+export * from './errors/CustomError';
+// TODO export * from './errors/HttpException';
+export * from './errors/RequestValidationError';
+
+export * from './middlewares/HandleFourOhFour';
+export * from './middlewares/ValidateRequest';
