+++ packages/common/src/middlewares/HandleFourOhFour.ts
@@ -1,0 +1,5 @@
+import { Request, Response, NextFunction } from 'express';
+
+export const handleFourOhFour = (req: Request, res: Response, next: NextFunction) => {
+  res.status(404).send('API not found');
+};
