+++ packages/common/src/middlewares/ValidateRequest.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import { validationResult } from 'express-validator';
-import { RequestValidationError } from '../common/RequestValidationError';
+import { RequestValidationError } from '../errors/RequestValidationError';
 
 export const validateRequest = (req: Request, res: Response, next: NextFunction) => {
   const errors = validationResult(req);
